+++ src/main/java/com/github/klousiaj/junit/DockerRuleParams.java
@@ -1,12 +1,15 @@
 package com.github.klousiaj.junit;
 
-public class DockerRuleParams {
+class DockerRuleParams {
 
   String imageName;
 
   String[] ports;
   String[] envs;
   String cmd;
+
+  boolean useRunning = false;
+  boolean leaveRunning = false;
 
   String portToWaitOn;
   public int waitTimeout;