+++ src/test/java/com/github/klousiaj/junit/RabbitIntegrationTest.java
@@ -1,6 +1,5 @@
-package integration;
+package com.github.klousiaj.junit;
 
-import com.github.klousiaj.junit.DockerRule;
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
 import org.junit.Assert;
@@ -19,6 +18,7 @@ public class RabbitIntegrationTest {
       .envs("RABBITMQ_DEFAULT_PASS=password1234")
 //      .waitForPort("5672/tcp")
       .waitForLog("Server startup complete")
+      .leaveRunning(true)
       .build();
 
   @Test
@@ -37,5 +37,21 @@ public class RabbitIntegrationTest {
     Connection conn = factory.newConnection();
     Map<String, Object> serverProperties = conn.getServerProperties();
     Assert.assertNotNull(serverProperties);
+  }
+
+  @Test
+  public void testContainerStaysRunning() throws Exception {
+    // run the after on the Rule. This should shutdown the container
+    // unless the leaveRunning flag is set.
+    rabbitRule.after();
+
+    Assert.assertNotEquals(-1, rabbitRule.getHostPort("5672/tcp"));
+    Assert.assertEquals(32779, rabbitRule.getHostPort("32779/tcp"));
+    Assert.assertEquals(32880, rabbitRule.getHostPort("5671/tcp"));
+    Assert.assertEquals(-1, rabbitRule.getHostPort("5675/tcp"));
+
+    // set the leaveRunning to false to make sure the container
+    // does actually get cleaned up.
+    rabbitRule.params.leaveRunning = false;
   }
 }